pyproject.toml
[project]
name = "atbatwatch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "click>=8.0",
    "questionary>=2.1.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "alembic>=1.13",
    "python-dotenv>=1.0",
]

[project.scripts]
atbatwatch = "atbatwatch.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["atbatwatch"]

[dependency-groups]
dev = [
    "ruff",
    "pyrefly",
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "fakeredis>=2.20",
    "aiosqlite>=0.20",
    "memray>=1.19.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pyrefly]
